Analysis renders automatically after updating entry content
