+++ app/(dashboard)/journal/[id]/page.tsx
@@ -20,43 +20,11 @@ const getEntry = async (id) => {
 
 const EntryPage = async ({ params }) => {
     const entry = await getEntry(params.id)
-    const { mood, summary, color, subject, negative } = entry?.analysis
-
-    const analysisData = [
-        { name: "Summary", value: summary },
-        { name: "Subject", value: subject },
-        { name: "Mood", value: mood },
-        { name: "Color", value: color },
-        { name: "Negative", value: negative ? "True" : "False" },
-    ]
 
     return (
-        <div className="h-full w-full grid grid-cols-3">
+        <div className="h-full w-full">
             <div className="col-span-2">
                 <Editor entry={entry} />
-            </div>
-
-            <div className="border-l border-black/10">
-                <div className="px-6 py-10" style={{ backgroundColor: color }}>
-                    <h2 className="text-2xl">Analysis</h2>
-                </div>
-                <div className="">
-                    <ul>
-                        {analysisData.map((item) => (
-                            <li
-                                key={item.name}
-                                className="px-2 py-4 flex items-center justify-between border-b border-t border-black/10"
-                            >
-                                <span className="text-lg font-semibold">
-                                    {item.name}
-                                </span>
-                                <span className="text-lg font-semibold">
-                                    {item.value}
-                                </span>
-                            </li>
-                        ))}
-                    </ul>
-                </div>
             </div>
         </div>
     )
