+++ app/api/journal/[id]/route.ts
@@ -1,6 +1,7 @@
 import { analyze } from "@/utils/ai"
 import { getUserByClerkId } from "@/utils/auth"
 import { prisma } from "@/utils/db"
+import { revalidatePath } from "next/cache"
 import { NextResponse } from "next/server"
 
 export const PATCH = async (request: Request, { params }) => {
@@ -30,7 +31,8 @@ export const PATCH = async (request: Request, { params }) => {
         },
         update: analysis,
     })
-    console.log("New Analysis?", updatedAnalysis)
 
-    return NextResponse.json({ data: updatedEntry })
+    return NextResponse.json({
+        data: { ...updatedEntry, analysis: updatedAnalysis },
+    })
 }
