+++ components/Editor.tsx
@@ -8,22 +8,60 @@ const Editor = ({ entry }) => {
     const [value, setValue] = useState(entry.content)
     const [isLoading, setIsLoading] = useState(false)
 
+    const [analysis, setAnalysis] = useState(entry.analysis)
+    const { mood, summary, color, subject, negative } = analysis
+    const analysisData = [
+        { name: "Summary", value: summary },
+        { name: "Subject", value: subject },
+        { name: "Mood", value: mood },
+        { name: "Color", value: color },
+        { name: "Negative", value: negative ? "True" : "False" },
+    ]
+
     useAutosave({
         data: value,
         onSave: async (_value) => {
             setIsLoading(true)
-            const updated = await updateEntry(entry.id, _value)
+            const data = await updateEntry(entry.id, _value)
+            setAnalysis(data.analysis)
             setIsLoading(false)
         },
     })
+
     return (
-        <div className="h-full w-full">
-            {isLoading && <>Loading...</>}
-            <textarea
-                className="w-full h-full p-8 text-xl outline-none"
-                value={value}
-                onChange={(e) => setValue(e.target.value)}
-            ></textarea>
+        <div className="h-full w-full grid grid-cols-3">
+            <div className="col-span-2">
+                {isLoading && <>Loading...</>}
+                <textarea
+                    className="w-full h-full p-8 text-xl outline-none"
+                    value={value}
+                    onChange={(e) => setValue(e.target.value)}
+                />
+            </div>
+
+            <div className="border-l border-black/10">
+                <div className="px-6 py-10" style={{ backgroundColor: color }}>
+                    <h2 className="text-2xl">Analysis</h2>
+                </div>
+
+                <div>
+                    <ul>
+                        {analysisData.map((item) => (
+                            <li
+                                key={item.name}
+                                className="px-2 py-4 flex items-center justify-between border-b border-t border-black/10"
+                            >
+                                <span className="text-lg font-semibold">
+                                    {item.name}
+                                </span>
+                                <span className="text-lg font-semibold">
+                                    {item.value}
+                                </span>
+                            </li>
+                        ))}
+                    </ul>
+                </div>
+            </div>
         </div>
     )
 }
